==> 04_PathUsage.ipynb <==
import pickle
import networkx as nx
import pandas as pd
import numpy as np
import time

with open(r'graph_centrality_codes/nodes_edges_weighted.pickle', 'rb') as handle:
    B_matrix_weighted, node_coordinates_weighted = pickle.load(handle)
    
with open(r'graph_centrality_codes/distance.pickle', 'rb') as handle:
    distance_array = pickle.load(handle)  
    
#distance_array (1st col - distance (m), 2nd col - time (s), 3rd col - time in traffic (s))

with open(r"script_output.pickle","rb") as handle:
    origin_destination_path_list,no_connection_list,origin_destination_path_df,no_connection_df=pickle.load(handle)

# origin_destination_path_df

# 0: origin_id
# 1: destination_id
# 2: path
# 3: S000_adjusted
# 4: SA01_adjusted
# 5: SA02_adjusted
# 6: SA03_adjusted
# 7: SE01_adjusted
# 8: SE02_adjusted
# 9: SE03_adjusted
# 10: SI01_adjusted
# 11: SI02_adjusted
# 12: SI03_adjusted

print('Number of rows in df: ' + str(len(origin_destination_path_list)))

# Make a copy of the origin_destination_path_list and origin_destination_path_df

origin_destination_path_list_copy = origin_destination_path_list.copy()
origin_destination_path_df_copy = origin_destination_path_df.copy()

# Subset of origin_destination_path_list and origin_destination_path_df

origin_destination_path_list_copy = origin_destination_path_list[0:1000]
origin_destination_path_df_copy = origin_destination_path_df[0:1000]

origin_destination_path_list_dummy=origin_destination_path_df_copy[2].tolist()
   
selected_columns = origin_destination_path_df_copy[[3,4,5,6,7,8,9,10,11,12]].values    

B_matrix_add=np.zeros([70983,10],dtype=float)

for i in range(len(origin_destination_path_list_dummy)):
    print(i)
    path=origin_destination_path_list_dummy[i]
    for j in range(len(path)-1):
        start_node=path[j]
        end_node=path[j+1]
        
        idx_array = np.where((B_matrix_weighted[:, 0] == start_node) & (B_matrix_weighted[:, 1] == end_node))
        if idx_array[0].size > 0:
            idx = idx_array[0][0]
        else:
            idx_array = np.where((B_matrix_weighted[:, 0] == end_node) & (B_matrix_weighted[:, 1] == start_node))
            if idx_array[0].size > 0:
                idx = idx_array[0][0]
            else:
                idx = None

        if idx is not None:
            B_matrix_add[idx,:] += selected_columns[i,:]

len(origin_destination_path_list_dummy)

origin_destination_path_list_dummy=origin_destination_path_df_copy[2].tolist()
   
selected_columns = origin_destination_path_df_copy[[3,4,5,6,7,8,9,10,11,12]].values    

B_matrix_add=np.zeros([70983,10],dtype=float)

for i in range(len(origin_destination_path_list_dummy)):
    print(i)
    path=origin_destination_path_list_dummy[i]
    for j in range(len(path)-1):
        start_node=path[j]
        end_node=path[j+1]
        try:
            idx = np.where((B_matrix_weighted[:, 0] == start_node) & (B_matrix_weighted[:, 1] == end_node))[0][0]
        except:
            idx = np.where((B_matrix_weighted[:, 0] == end_node) & (B_matrix_weighted[:, 1] == start_node))[0][0]
        B_matrix_add[idx,:]+=selected_columns[i,:]

print(sum(B_matrix_add[:, 6] != 0))

# Count how many elements there are total in the paths of the origin_destination_path_list_copy

total_elements_copy = 0

for i in range(len(origin_destination_path_list_copy)):
    total_elements_copy += len(origin_destination_path_list_copy[i][2])


# Count how many elements there are total in the paths of the origin_destination_path_list

total_elements = 0

for i in range(len(origin_destination_path_list)):
    total_elements += len(origin_destination_path_list[i][2])


print('Total elements in the paths of the origin_destination_path_list_copy: ' + str(total_elements_copy))
print('Total elements in the paths of the origin_destination_path_list: ' + str(total_elements))

ratio = total_elements / total_elements_copy

# Rename columns

origin_destination_path_df_copy.columns = ['origin_id', 'destination_id', 'path', 'S000_adjusted', 'SA01_adjusted', 'SA02_adjusted', 'SA03_adjusted', 'SE01_adjusted', 'SE02_adjusted', 'SE03_adjusted', 'SI01_adjusted', 'SI02_adjusted', 'SI03_adjusted']

pairlist_actual=[]

for j in range(len(origin_destination_path_list_copy)):
    path = origin_destination_path_list_copy[j][2]
    for i in range(len(path) - 1):
        pair=(min([path[i],path[i+1]]),max([path[i],path[i+1]]))
        if pair not in pairlist_actual:
            pairlist_actual.append(pair)

print(len(pairlist_actual))

# Initialize array of 0s with 10 columns and add to B_matrix_weighted
zero_columns = np.zeros((B_matrix_weighted.shape[0], 10))
B_matrix_weighted_array = np.hstack((B_matrix_weighted, zero_columns))
B_matrix_weighted_array = B_matrix_weighted_array.astype(float)

B_matrix_weighted.shape

B_matrix_weighted_array.shape[0]

print(B_matrix_weighted_array.shape)

# B_matrix_weighted_df = pd.DataFrame(B_matrix_weighted_array)
# B_matrix_weighted_df

# Approach 1: Indexing the original B_matrix_weighted in array form (Linear, very slow)

start=time.time()
pairlist_test_1=[]

for index, row in origin_destination_path_df_copy.iterrows():
    path = row['path']
    for i in range(len(path) - 1):
        
        idx = np.where((B_matrix_weighted_array[:, 0] == path[i]) & (B_matrix_weighted_array[:, 1] == path[i+1]))[0]

        if len(idx) > 0:
            B_matrix_weighted_array[idx[0], 6] = B_matrix_weighted_array[idx[0], 6] + row['S000_adjusted']
            B_matrix_weighted_array[idx[0], 7] = B_matrix_weighted_array[idx[0], 7] + row['SA01_adjusted']
            B_matrix_weighted_array[idx[0], 8] = B_matrix_weighted_array[idx[0], 8] + row['SA02_adjusted']
            B_matrix_weighted_array[idx[0], 9] = B_matrix_weighted_array[idx[0], 9] + row['SA03_adjusted']
            B_matrix_weighted_array[idx[0], 10] = B_matrix_weighted_array[idx[0], 10] + row['SE01_adjusted']
            B_matrix_weighted_array[idx[0], 11] = B_matrix_weighted_array[idx[0], 11] + row['SE02_adjusted']
            B_matrix_weighted_array[idx[0], 12] = B_matrix_weighted_array[idx[0], 12] + row['SE03_adjusted']
            B_matrix_weighted_array[idx[0], 13] = B_matrix_weighted_array[idx[0], 13] + row['SI01_adjusted']
            B_matrix_weighted_array[idx[0], 14] = B_matrix_weighted_array[idx[0], 14] + row['SI02_adjusted']
            B_matrix_weighted_array[idx[0], 15] = B_matrix_weighted_array[idx[0], 15] + row['SI03_adjusted']
        elif len(idx) == 0:
            idx = np.where((B_matrix_weighted_array[:, 1] == path[i]) & (B_matrix_weighted_array[:, 0] == path[i+1]))[0]
            if len(idx) > 0:
                B_matrix_weighted_array[idx[0], 6] = B_matrix_weighted_array[idx[0], 6] + row['S000_adjusted']
                B_matrix_weighted_array[idx[0], 7] = B_matrix_weighted_array[idx[0], 7] + row['SA01_adjusted']
                B_matrix_weighted_array[idx[0], 8] = B_matrix_weighted_array[idx[0], 8] + row['SA02_adjusted']
                B_matrix_weighted_array[idx[0], 9] = B_matrix_weighted_array[idx[0], 9] + row['SA03_adjusted']
                B_matrix_weighted_array[idx[0], 10] = B_matrix_weighted_array[idx[0], 10] + row['SE01_adjusted']
                B_matrix_weighted_array[idx[0], 11] = B_matrix_weighted_array[idx[0], 11] + row['SE02_adjusted']
                B_matrix_weighted_array[idx[0], 12] = B_matrix_weighted_array[idx[0], 12] + row['SE03_adjusted']
                B_matrix_weighted_array[idx[0], 13] = B_matrix_weighted_array[idx[0], 13] + row['SI01_adjusted']
                B_matrix_weighted_array[idx[0], 14] = B_matrix_weighted_array[idx[0], 14] + row['SI02_adjusted']
                B_matrix_weighted_array[idx[0], 15] = B_matrix_weighted_array[idx[0], 15] + row['SI03_adjusted']
        if idx[0]not in pairlist_test_1:
            pairlist_test_1.append(idx[0])

end=time.time()

print(end-start)
print(len(pairlist_test_1))

print(len(np.unique(pairlist_test_1)))

print(sum(B_matrix_weighted_array[:, 6] != 0))

B_matrix_weighted_dict = {(row[0], row[1]): row for row in B_matrix_weighted_array}

len(B_matrix_weighted_dict)

unique_rows = np.unique(B_matrix_weighted_array, axis=0)
num_unique_rows = unique_rows.shape[0]
num_total_rows = B_matrix_weighted_array.shape[0]

if num_unique_rows < num_total_rows:
    print("There are duplicate rows in B_matrix_weighted_array.")
else:
    print("There are no duplicate rows in B_matrix_weighted_array.")

unique_rows, counts = np.unique(B_matrix_weighted_array, axis=0, return_counts=True)
duplicate_rows = unique_rows[counts > 1]

print(len(duplicate_rows))

# Create a new array that only contains the first two elements of each row
first_two_columns = B_matrix_weighted_array[:, :2]

# Find duplicate rows in the new array
unique_rows, counts = np.unique(first_two_columns, axis=0, return_counts=True)
duplicate_rows = unique_rows[counts > 1]

print(len(duplicate_rows))

print(len(unique_rows))

# put unique_rows and counts into a single dataframe

unique_rows_df = pd.DataFrame(unique_rows)
counts_df = pd.DataFrame(counts)

unique_rows_counts_df = pd.concat([unique_rows_df, counts_df], axis=1)



unique_rows_counts_df

# Approach 2: Converting B_matrix_weighted to a dictionary for faster lookups (O(1) lookups, faster)

# Convert B_matrix_weighted_array to a dictionary for faster lookups
B_matrix_weighted_dict = {(row[0], row[1]): row for row in B_matrix_weighted_array}

start = time.time()

for _, row in origin_destination_path_df_copy.iterrows():
    path = row['path']
    for i in range(len(path) - 1):
        pair = (path[i], path[i+1])
        reverse_pair = (path[i+1], path[i])

        if pair in B_matrix_weighted_dict:
            B_matrix_weighted_dict[pair][6:16] += row['S000_adjusted':'SI03_adjusted']
        elif reverse_pair in B_matrix_weighted_dict:
            B_matrix_weighted_dict[reverse_pair][6:16] += row['S000_adjusted':'SI03_adjusted']

# Convert the dictionary back to a numpy array
B_matrix_weighted_array = np.array(list(B_matrix_weighted_dict.values()))

end = time.time()

print(end - start)

len(B_matrix_weighted_array[B_matrix_weighted_array[:, 6] != 0])

# Estimated time to run the entire origin_destination_path_df

test_time = (end - start) * ratio

print(test_time / 60 / 60 / 24)

# # Approach 2: Indexing the original B_matrix_weighted in dataframe form (This would be slower than the first approach but is working)

# This iterates over each row of the origin_destination_path_df dataframe and for each row, it iterates over each pair of adjacent nodes in the path column.
# For each pair of nodes, it checks if the nodes are present in the first and second columns (or reversed) of the B_matrix_weighted_df dataframe.
# If the nodes are present, it adds the S000_adjusted value to the 6th column of the B_matrix_weighted_df dataframe.

# start=time.time()

# for index, row in origin_destination_path_df.iterrows():
#     for i in range(len(row['path']) - 1):
#         if (row['path'][i] in B_matrix_weighted_df[0] and row['path'][i + 1] in B_matrix_weighted_df[1]):
#             B_matrix_weighted_df.loc[(B_matrix_weighted_df[0] == row['path'][i]) & (B_matrix_weighted_df[1] == row['path'][i + 1]), 6] += row['S000_adjusted']
#             # print(B_matrix_weighted_df.loc[(B_matrix_weighted_df[0] == row['path'][i]) & (B_matrix_weighted_df[1] == row['path'][i + 1])])
#         elif (row['path'][i+1] in B_matrix_weighted_df[0] and row['path'][i] in B_matrix_weighted_df[1]):
#             B_matrix_weighted_df.loc[(B_matrix_weighted_df[1] == row['path'][i]) & (B_matrix_weighted_df[0] == row['path'][i + 1]), 6] += row['S000_adjusted']
#             # print(B_matrix_weighted_df.loc[(B_matrix_weighted_df[0] == row['path'][i]) & (B_matrix_weighted_df[1] == row['path'][i + 1])])

# end=time.time()

# print(end-start)

# print(len(B_matrix_weighted_df[B_matrix_weighted_df[6] != 0]))

# B_matrix_weighted_df

print('Number of node pairs updated in B_matrix_weighted: ' + str(len(B_matrix_weighted_array[B_matrix_weighted_array[:, 6] != 0])))

# Convert B_matrix_weighted_df back to a numpy array

# B_matrix_weighted_updated = B_matrix_weighted_df.to_numpy()

# Export B_matrix_weighted_updated as a pickle file

with open('B_matrix_weighted_updated.pickle', 'wb') as handle:
    pickle.dump(B_matrix_weighted_array, handle, protocol=pickle.HIGHEST_PROTOCOL)
